*= blade modes consolidation, ++ blade_mode_k, mode_a kurtosis The consolidation of blade modes is a step towards dynamic blade instantiation ++ config_input_shape.h single place for static input dimension sizes changes to config_input_shape (I think) does cause long compilation times... which typically happens between blade modes...

+++ src/hpguppi_xgpu_databuf.h
@@ -21,7 +21,7 @@
 #include "hpguppi_databuf.h"
 #include "uvh5.h"
 
-#define XGPU_BLOCK_DATA_SIZE 2*BLOCK_DATA_SIZE // in bytes, from guppi_daq_server
+#define XGPU_BLOCK_DATA_SIZE ((size_t)2*BLOCK_DATA_SIZE) // in bytes, from guppi_daq_server
 
 #define XGPU_INTEGRATE_AS_CF64_ON_CPU
 #define BLADE_CORRELATOR
